refactor: remove useless index to keep default export
